Add Melody class
A Melody has a Mode, a Voice Range, and a list of Notes and the intervals
between successive Notes. Checks if the Pitch of an added Note is within
the bounds of the Voice Range. Whether a Pitch conforms with the Mode is
NOT checked (yet), as occasional sharpened/flattened pitches (most
often the leading tone) will occur.