Add VoiceType to easily get adjacent Voice Ranges
The values of the voice types are integers ascending from the voice type
with the highest pitch range to the lowest. This makes it easy to get
the VoiceTypes that are adjacent in pitch range to a given VoiceType (a
Counterpoint should harmonize a CantusFirmus in an adjacent pitch
range).

Each VoiceType has a reference to a VoiceRange that corresponds to the
VoiceType's name. Each Melody has a VoiceType through which it can
access its own pitch range and adjacent ranges.